Rename bridge.py → kernelagent_bridge.py
